requirements.txt
aiohttp==3.9.1
discord.py==2.3.2
python-dotenv==1.0.0
google-api-python-client==2.111.0
google-auth==2.25.2
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
